+++ tools/wsq/CLI/wsq.cpp
@@ -50,6 +50,27 @@ namespace {
 	}
 }
 
+namespace {
+    void Display_(const websck::dFields &Fields) {
+        sdr::sRow Row = Fields.First();
+
+        while ( Row != qNIL ) {
+            cio::COut << Fields(Row).Label << " -> " << Fields(Row).Value << txf::nl;
+
+            Row = Fields.Next(Row);
+        }
+
+        cio::COut.Commit();
+    }
+
+    void Display_(const websck::dHeader &Header) {
+        cio::COut << Header.FirstLine << txf::nl;
+
+        Display_(Header.Fields);
+
+        cio::COut.Commit();
+    }
+}
 #define TEXT "My name is JohnLorem ipsum dolor sit amet, consectetur adipiscing elit. Sed non risus. Suspendisse lectus tortor, dignissim sit amet, adipiscing nec, ultricies sed, dolor. Cras elementum ultrices diam. Maecenas ligula massa, varius a, semper congue, euismod non, mi. Proin porttitor, orci nec nonummy molestie, enim est eleifend mi, non fermentum diam nisl sit amet erat. Duis semper. Duis arcu massa, scelerisque vitae, consequat in, pretium a, enim. Pellentesque congue. Ut in risus volutpat libero pharetra tempor. Cras vestibulum bibendum augue. Praesent egestas leo in pede. Praesent blandit odio eu enim. Pellentesque sed dui ut augue blandit sodales. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Aliquam nibh. Mauris ac mauris sed pede pellentesque fermentum. Maecenas adipiscing ante non diam sodales hendrerit"
 
 namespace {
@@ -63,10 +84,18 @@ namespace {
             fdr::rRWDriver *Driver,
             const ntvstr::char__ *Origin ) override
         {
+        qRH
+            websck::wHeader Header;
+        qRB
             cio::COut << "New connection !!!" << txf::nl << txf::commit;
 
-            websck::Handshake(*Driver);
+            Header.Init();
+            websck::Handshake(*Driver, Header);
 
+            Display_(Header);
+        qRR
+        qRT
+        qRE
             return NULL;
         }
         csdscb::action__ CSDSCBProcess(